+++ test/elysia.test.ts
@@ -89,4 +89,18 @@ describe('Elysia', () => {
 			message: 'Hello, world!'
 		})
 	})
+
+	it('handle config.strictPath', async () => {
+		const loose = new Elysia().group('/a', (app) => app.get('/', () => 'Hi'))
+
+		expect(await loose.handle(req('/a')).then(x => x.status)).toBe(200)
+		expect(await loose.handle(req('/a/')).then(x => x.status)).toBe(200)
+
+		const strict = new Elysia({
+			strictPath: true
+		}).group('/a', (app) => app.get('/', () => 'Hi'))
+
+		expect(await strict.handle(req('/a')).then(x => x.status)).toBe(404)
+		expect(await strict.handle(req('/a/')).then(x => x.status)).toBe(200)
+	})
 })